requirements.txt
# App
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.
python-decouple==3.4
gunicorn==20.0.4
psycopg2-binary==2.8.6
django-test-without-migrations==0.6
django_nose==1.4.7
model_bakery==1.2.1
django-cors-headers==3.7.0
django_extensions==3.1.1
boto3==1.17.39
django-storages==1.11.1
pytest==6.2.3
pytest-django==4.1.0

# Heroku
dj-database-url==0.5.0
dj-static==0.0.6